refactor intermediate bindings
